fix location and var names in class
